requirements.txt
asn1crypto==0.24.0
astroid==2.3.1
cffi==1.12.3
colorama==0.4.1
cryptography==2.7
Django==2.2.5
django-rest-knox==4.1.0
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pycparser==2.19
pylint==2.4.2
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
